Switch from South to Django migrations
