+++ docs/releases.rst
@@ -25,11 +25,18 @@ This is a stable release supporting Django 1.8 and Python 3.
 - Updated sample project to work with Django 1.8 and django-mptt 0.7 (#25)
 - Fixed bug that prevented deletion of items from admin changelist (#54)
 - Fixed bug where reordering items in the admin would disorder tree (#42)
+- Switched to Django migrations (#63)
 
 Backwards Incompatible Changes
-___________________________________
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - ``MenuItem`` model no longer uses ``.tree`` for ``TreeManager`` methods. Instead use ``.objects`` (change from django-mptt 0.5)
+- If you are upgrading from 0.9.2 (or earlier) and are currently using Django 1.8, then you will need to
+  run ``python manage.py migrate --fake-initial`` since we have converted from using South to Django
+  migrations. Django 1.7 does "fake-initial" behavior by default.
+- The ``post_save`` signal which updates MenuItems when their corresponding object changes will NOT
+  be active during data migrations. If you update any objects during a data migration, and that
+  object has a corresponding MenuItem, you will need to manually update the HREF of the MenuItem.
 
 v0.9.2 (Released 2015-09-02)
 ------------------------------------
@@ -52,7 +59,7 @@ considered a release candidate for an offical v1.0.
 - Added tox support for testing
 
 Backwards Incompatible Changes
-___________________________________
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - ``treenav_refresh_hrefs`` and ``treenav_clean_cache`` were moved under the admin and require staff access
 - Including url patterns has been simplified and ``treenav.urls.admin`` were removed
