+++ runtests.py
@@ -60,7 +60,9 @@ if not settings.configured:
                 },
             },
         ],
-
+        # Load models directly to pick up test-only models
+        # See: http://stackoverflow.com/a/25267435/347942
+        MIGRATION_MODULES={'treenav': 'treenav.migrations_not_run_during_tests'},
         SECRET_KEY='this-is-just-for-tests-so-not-that-secret',
         ROOT_URLCONF='treenav.tests.urls',
     )
