+++ treenav/__init__.py
@@ -1,1 +1,2 @@
 __version__ = '0.9.2'
+default_app_config = 'treenav.apps.TreenavConfig'
