+++ treenav/models.py
@@ -4,7 +4,6 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes import fields
-from django.db.models.signals import post_save
 from django.core.cache import cache
 from django.core.urlresolvers import reverse
 
@@ -176,28 +175,3 @@ class MenuItem(MPTTModel):
 
     def __unicode__(self):
         return self.slug
-
-
-def treenav_save_other_object_handler(sender, instance, created, **kwargs):
-    """
-    This signal attempts to update the HREF of any menu items that point to
-    another model object, when that objects is saved.
-    """
-    cache_key = 'django-treenav-menumodels'
-    if sender == MenuItem:
-        cache.delete(cache_key)
-    menu_models = cache.get(cache_key)
-    if not menu_models:
-        menu_models = []
-        for menu_item in MenuItem.objects.exclude(content_type__isnull=True):
-            menu_models.append(menu_item.content_type.model_class())
-        cache.set(cache_key, menu_models)
-    # only attempt to update MenuItem if sender is known to be referenced
-    if sender in menu_models:
-        ct = ContentType.objects.get_for_model(sender)
-        items = MenuItem.objects.filter(content_type=ct, object_id=instance.pk)
-        for item in items:
-            if item.href != instance.get_absolute_url():
-                item.href = instance.get_absolute_url()
-                item.save()
-post_save.connect(treenav_save_other_object_handler)
